# CSC RU segment algorithm tuned for ME0