… RecoLocalMuon/GEMSegment/plugins/ME0SegmentAlgorithm.h
… RecoLocalMuon/GEMSegment/plugins/ME0SegmentAlgorithmBase.h
… RecoLocalMuon/GEMSegment/plugins/ME0SegmentBuilder.cc
… RecoLocalMuon/GEMSegment/plugins/SealModule.cc
… RecoLocalMuon/GEMSegment/python/ME0SegmentsAlgorithm_cfi.py
… RecoLocalMuon/GEMSegment/python/ME0SegmentsRU_cfi.py
… RecoLocalMuon/GEMSegment/python/me0Segments_cfi.py
… RecoLocalMuon/GEMSegment/test/TestME0SegmentAnalyzer.cc
… RecoLocalMuon/GEMSegment/test/testme0segementanalyzer_cfg.py