Translated CSC RU algorithm (https://indico.cern.ch/event/578871/contributions/2345178/attachments/1358089/2053858/16_10_19-Voytishin_Palichik-CSCSegBuilderCSCWeekly.pdf) for use in ME0. It is not a direct port as some of the assumptions made for the CSC do not apply to ME0. It is now set as the default. Also updated standard segment algorithm to use the chamber as the coordinate reference frame. 

Right now the default cfg parameters are for a 10 eta partition geometry and 768 strips...this will change once the geometry is finalized.
Automatically ported from CMSSW_9_0_X #17677 (original by @nickmccoll).
Please wait for a new IB (12 to 24H) before requesting to test this PR.