@cmsbuild 2017-03-02
A new Pull Request was created by  @cmsbuild  for master.

It involves the following packages:

RecoLocalMuon/GEMSegment


@cmsbuild, @cvuosalo, @slava77, @kpedro88, @davidlange6 can you please review it and eventually sign? Thanks.
@bellan, @jhgoh this is something you requested to watch as well.
@muzaffar, @davidlange6, @smuzaffar you are the release manager for this.

cms-bot commands are listed here cms-sw/cmssw#13028