@nickmccoll 2017-03-02
@slava77 Thanks for the info, do you want the 10k event memory profile to be with PU, or can that be w/o and just do the timing on a few events with PU? I am using pre4 GEN-SIM relvals to do this test, but the PU mixing is taking a while.