@slava77 2017-03-03
On 3/2/17 2:05 PM, Nick McColl wrote:
> @slava77 <https://github.com/slava77> Thanks for the info, do you want
> the 10k event memory profile to be with PU, or can that be w/o and just
> do the timing on a few events with PU? I am using pre4 GEN-SIM relvals
> to do this test, but the PU mixing is taking a while.

I think that you can use the "step2" outputs directly
and run just the ME or muon local reco.

RECO:me0LocalReco in the cmsDriver.

/RelValTenMuExtendedE_0_200/CMSSW_9_0_0_pre4-90X_upgrade2023_realistic_v3_2023D4Timing-v1/GEN-SIM-DIGI-RAW


/RelValTTbar_14TeV/CMSSW_9_0_0_pre4-PU25ns_90X_upgrade2023_realistic_v3_D4TPU200c2-v1/GEN-SIM-DIGI-RAW
.. c2 is apparently for two core jobs


> 
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/cms-sw/cmssw/pull/17687#issuecomment-283797480>, or
> mute the thread
> <https://github.com/notifications/unsubscribe-auth/AEdcblG-Edt95MgLj3oldQhJOT4VS2Hpks5rhz1BgaJpZM4MQ7Bw>.
>