@nickmccoll 2017-03-03
Thanks! I was under the impression that that dataset didn't include PU. Anyways:

I ran the following:
cmsDriver.py step3  --conditions auto:phase2_realistic -n 1000 --era Phase2C2_timing --eventcontent FEVTDEBUGHLT,MINIAODSIM,DQM --runUnscheduled  -s RECO:me0LocalReco --datatier GEN-SIM-RECO,MINIAODSIM,DQMIO --geometry Extended2023D4 --customise Validation/Performance/TimeMemoryInfo.py --filein das:/RelValTenMuExtendedE_0_200/CMSSW_9_0_0_pre4-90X_upgrade2023_realistic_v3_2023D4Timing-v1/GEN-SIM-DIGI-RAW

And here are the memory and time reports. Let me know if you want other information:

TimeReport ---------- Event  Summary ---[sec]----
TimeReport       event loop CPU/event = 1.831721
TimeReport      event loop Real/event = 1.963321
TimeReport     sum Streams Real/event = 1.837479
TimeReport efficiency CPU/Real/thread = 0.932971

TimeReport ---------- Path   Summary ---[Real sec]----
TimeReport  per event     per exec  Name
TimeReport   0.000000     0.000000  reconstruction_step
TimeReport  per event     per exec  Name

TimeReport -------End-Path   Summary ---[Real sec]----
TimeReport  per event     per exec  Name
TimeReport   1.836215     1.836215  FEVTDEBUGHLToutput_step
TimeReport   1.836390     1.836390  MINIAODSIMoutput_step
TimeReport   1.836435     1.836435  DQMoutput_step
TimeReport  per event     per exec  Name

TimeReport ---------- Modules in Path: reconstruction_step ---[Real sec]----
TimeReport  per event    per visit  Name
TimeReport  per event    per visit  Name

TimeReport ------ Modules in End-Path: FEVTDEBUGHLToutput_step ---[Real sec]----
TimeReport  per event    per visit  Name
TimeReport   0.215789     0.215789  FEVTDEBUGHLToutput

TimeReport ------ Modules in End-Path: MINIAODSIMoutput_step ---[Real sec]----
TimeReport  per event    per visit  Name
TimeReport   0.000115     0.000115  MINIAODSIMoutput

TimeReport ------ Modules in End-Path: DQMoutput_step ---[Real sec]----
TimeReport  per event    per visit  Name
TimeReport   0.000001     0.000001  DQMoutput
TimeReport  per event    per visit  Name

TimeReport ---------- Module Summary ---[Real sec]----
TimeReport  per event     per exec    per visit  Name
TimeReport   0.000001     0.000001     0.000001  DQMoutput
TimeReport   0.215789     0.215789     0.215789  FEVTDEBUGHLToutput
TimeReport   0.000115     0.000115     0.000115  MINIAODSIMoutput
TimeReport   0.000021     0.000021     0.000021  TriggerResults
TimeReport   0.009117     0.009117     0.009117  me0RecHits
TimeReport   1.518472     1.518472     1.518472  me0Segments
TimeReport  per event     per exec    per visit  Name

T---Report end!

MemoryReport> Peak virtual size 1677.17 Mbytes
 Key events increasing vsize: 
[51] run: 1 lumi: 164 event: 8152  vsize = 1032.17 deltaVsize = 17.75 rss = 484.504 delta = 18.793
[101] run: 1 lumi: 167 event: 8304  vsize = 1068.67 deltaVsize = 35.75 rss = 505.887 delta = 21.3828
[500] run: 1 lumi: 3 event: 106  vsize = 1513.17 deltaVsize = 420.75 rss = 892.242 delta = 386.355
[560] run: 1 lumi: 7 event: 310  vsize = 1621.17 deltaVsize = 56 rss = 1011.19 delta = 118.949
[598] run: 1 lumi: 7 event: 350  vsize = 1677.17 deltaVsize = 56 rss = 1038.46 delta = 27.2695
[600] run: 1 lumi: 14 event: 656  vsize = 1677.17 deltaVsize = 0 rss = 1013.2 delta = -25.2617
[599] run: 1 lumi: 7 event: 347  vsize = 1677.17 deltaVsize = 0 rss = 1012.71 delta = -25.75
[598] run: 1 lumi: 7 event: 350  vsize = 1677.17 deltaVsize = 56 rss = 1038.46 delta = 27.2695
TimeReport> Time report complete in 1963.32 seconds
 Time Summary: 
 - Min event:   1.11996
 - Max event:   9.91998
 - Avg event:   1.83661
 - Total loop:  1962.64
 - Total job:   1963.32
 Event Throughput: 0.509518 ev/s
 CPU Summary: 
 - Total loop:  1831
 - Total job:   1831.44