@nickmccoll 2017-03-04
**With PR, /RelValTTbar_14TeV/CMSSW_9_0_0_pre4-PU25ns_90X_upgrade2023_realistic_v3_D4TPU200c2-v1/GEN-SIM-DIGI-RAW**

MemoryReport> Peak virtual size 2677.7 Mbytes
 Key events increasing vsize: 
[4] run: 1 lumi: 97 event: 4804  vsize = 1762.88 deltaVsize = 2.28906 rss = 1072.25 delta = -0.542969
[3] run: 1 lumi: 97 event: 4803  vsize = 1760.59 deltaVsize = 48.1367 rss = 1072.79 delta = 30.7891
[5] run: 1 lumi: 97 event: 4802  vsize = 1922.88 deltaVsize = 160 rss = 1086.37 delta = 13.582
[371] run: 1 lumi: 146 event: 7271  vsize = 2468.95 deltaVsize = 160 rss = 1147.29 delta = 60.918
[882] run: 1 lumi: 65 event: 3233  vsize = 2677.7 deltaVsize = 112 rss = 1172.84 delta = 25.5508
[884] run: 1 lumi: 65 event: 3234  vsize = 2677.7 deltaVsize = 0 rss = 1150.89 delta = -21.9492
[883] run: 1 lumi: 65 event: 3231  vsize = 2677.7 deltaVsize = 0 rss = 1153.57 delta = -19.2734
[882] run: 1 lumi: 65 event: 3233  vsize = 2677.7 deltaVsize = 112 rss = 1172.84 delta = -3.94141
TimeReport> Time report complete in 10520 seconds
 Time Summary: 
 - Min event:   5.63124
 - Max event:   19.6288
 - Avg event:   8.21454
 - Total loop:  10518.6
 - Total job:   10520
 Event Throughput: 0.0950698 ev/s
 CPU Summary: 
 - Total loop:  8376.61
 - Total job:   8377.42



**Without PR, /RelValTTbar_14TeV/CMSSW_9_0_0_pre4-PU25ns_90X_upgrade2023_realistic_v3_D4TPU200c2-v1/GEN-SIM-DIGI-RAW**

TimeReport ---------- Event  Summary ---[sec]----
TimeReport       event loop CPU/event = 5.226798
TimeReport      event loop Real/event = 7.451737
TimeReport     sum Streams Real/event = 5.097233
TimeReport efficiency CPU/Real/thread = 0.701420

TimeReport ---------- Path   Summary ---[Real sec]----
TimeReport  per event     per exec  Name
TimeReport   0.000000     0.000000  reconstruction_step
TimeReport  per event     per exec  Name

TimeReport -------End-Path   Summary ---[Real sec]----
TimeReport  per event     per exec  Name
TimeReport   5.093572     5.093572  FEVTDEBUGHLToutput_step
TimeReport   5.093799     5.093799  MINIAODSIMoutput_step
TimeReport   5.093850     5.093850  DQMoutput_step
TimeReport  per event     per exec  Name

TimeReport ---------- Modules in Path: reconstruction_step ---[Real sec]----
TimeReport  per event    per visit  Name
TimeReport  per event    per visit  Name

TimeReport ------ Modules in End-Path: FEVTDEBUGHLToutput_step ---[Real sec]----
TimeReport  per event    per visit  Name
TimeReport   0.093840     0.093840  FEVTDEBUGHLToutput

TimeReport ------ Modules in End-Path: MINIAODSIMoutput_step ---[Real sec]----
TimeReport  per event    per visit  Name
TimeReport   0.000160     0.000160  MINIAODSIMoutput

TimeReport ------ Modules in End-Path: DQMoutput_step ---[Real sec]----
TimeReport  per event    per visit  Name
TimeReport   0.000001     0.000001  DQMoutput
TimeReport  per event    per visit  Name

TimeReport ---------- Module Summary ---[Real sec]----
TimeReport  per event     per exec    per visit  Name
TimeReport   0.000001     0.000001     0.000001  DQMoutput
TimeReport   0.093840     0.093840     0.093840  FEVTDEBUGHLToutput
TimeReport   0.000160     0.000160     0.000160  MINIAODSIMoutput
TimeReport   0.000031     0.000031     0.000031  TriggerResults
TimeReport   0.008767     0.008767     0.008767  me0RecHits
TimeReport   0.005204     0.005204     0.005204  me0Segments
TimeReport  per event     per exec    per visit  Name

T---Report end!

MemoryReport> Peak virtual size 3317.64 Mbytes
 Key events increasing vsize: 
[4] run: 1 lumi: 97 event: 4804  vsize = 1762.82 deltaVsize = 2.28906 rss = 1073.93 delta = -0.253906
[3] run: 1 lumi: 97 event: 4803  vsize = 1760.53 deltaVsize = 48.1367 rss = 1074.18 delta = 31.9453
[5] run: 1 lumi: 97 event: 4802  vsize = 1922.82 deltaVsize = 160 rss = 1088.6 delta = 14.4141
[371] run: 1 lumi: 146 event: 7271  vsize = 2565.14 deltaVsize = 160 rss = 1321.27 delta = 232.676
[724] run: 1 lumi: 96 event: 4773  vsize = 3093.64 deltaVsize = 160 rss = 1490.92 delta = 169.645
[986] run: 1 lumi: 76 event: 3787  vsize = 3317.64 deltaVsize = 0 rss = 1601.72 delta = -42.0156
[985] run: 1 lumi: 76 event: 3783  vsize = 3317.64 deltaVsize = 0 rss = 1586.14 delta = -57.6016
[984] run: 1 lumi: 76 event: 3784  vsize = 3317.64 deltaVsize = 112 rss = 1643.74 delta = 58.2656
TimeReport> Time report complete in 7451.74 seconds
 Time Summary: 
 - Min event:   3.54304
 - Max event:   24.5769
 - Avg event:   5.09408
 - Total loop:  7450.75
 - Total job:   7451.74
 Event Throughput: 0.134215 ev/s
 CPU Summary: 
 - Total loop:  5225.38
 - Total job:   5226.01